% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIRE data set.}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{miles from fire station}
  \item{DAMAGE}{cost of damage in 1000 dollars per unit}
}
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
A data set collected examining cost of damage from house fires
}
\keyword{datasets}
