% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial simulation function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of iterations to be performed}

\item{n}{Number of trials in each iteration}

\item{p}{Probability of success}
}
\value{
A barplot of the proportions of sucesses
}
\description{
This function creates a bar plot of the proportions of successes in a binomial experiment
over the given number of iterations with the sample size and probability provided.
}
\examples{
mybin(iter=1000,n=18, p=0.3)
}
