% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{Bootstrapping function}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{iter}{Number of iterations to be performed}

\item{x}{The vector containing values to do operations on}

\item{fun}{Name of the function to use when calculating confidence intervals}

\item{alpha}{Dictates that we are creating confidence intervals of size 100(1-alpha)%}
}
\value{
A list object featuring useful information from the calculations
}
\description{
This function creates a bar plot from a sample of values using a specified
function to calculate statistics and builds confidence intervals of a specified size
}
\examples{
myboot(iter=1000,x=sample,fun="var",alpha=0.2)
}
