% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Confidence interval function}
\usage{
myci(x)
}
\arguments{
\item{x}{The vector containing data to calculate the ci}
}
\value{
A 95 percent confidence interval for the given data set
}
\description{
This function uses a given set of values from a single sample
to calculate a 95 percent confidence interval for mean (mu)
}
\examples{
myci(rnorm(30,mean=10,sd=12))
}
