% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlike.R
\name{mymaxlike}
\alias{mymaxlike}
\title{Max likelihood function}
\usage{
mymaxlike(lfun, x, param, ...)
}
\arguments{
\item{lfun}{Name of the function to use when calculating log likelihood}

\item{x}{The vector containing values to do operations on}

\item{param}{parameter values to be used for the variable we want to examine (probability, lambda, etc.)}
}
\value{
A list object featuring useful information from the calculations
}
\description{
This function uses inputs to create a plot of the maximum likelihood
distribution for a given set of parameters using the function provided
}
\examples{
mymaxlike(x=c(3,3,4,3,4,5,5,4),param=seq(0,1,length=1000),lfun=logbin)
}
